% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nts_build.R
\name{build_nts_csv}
\alias{build_nts_csv}
\title{Build the National Travel Survey data set from a CSV input}
\usage{
build_nts_csv(nts_dir, csv_dir = NULL, nts_names = nts_names)
}
\arguments{
\item{nts_dir}{Directory to folder where NTS .tab data files are stored}

\item{csv_dir}{Directory to input csv file}
}
\value{
A tibble of NTS tables joined together
}
\description{
Build the National Travel Survey data set from a CSV input
}
