% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nts_build.R
\name{build_nts_list}
\alias{build_nts_list}
\title{Build the National Travel Survey data set from a list}
\usage{
build_nts_list(nts_dir, nts_extract = "Default")
}
\arguments{
\item{nts_extract}{}
}
\description{
Build the National Travel Survey data set from a list
}
