% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nts_read_tab}
\alias{nts_read_tab}
\title{Read in individual NTS data tables}
\usage{
nts_read_tab(df_name, df_cols = NULL, nts_dir)
}
\arguments{
\item{df_name}{A data table name to read}

\item{nts_dir}{Path to folder .tab files are in}

\item{year_min}{Minimum year of data}

\item{year_max}{Maximum year of data}
}
\value{
A tibble of an NTS data table
}
\description{
Read in individual NTS data tables
}
