% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nts_build.R
\name{build_nts_list}
\alias{build_nts_list}
\title{Build the National Travel Survey data set from a list}
\usage{
build_nts_list(nts_dir, nts_extract = "Default", nts_names = nts_names)
}
\arguments{
\item{nts_dir}{Directory to folder where NTS .tab data files are stored}

\item{nts_extract}{Default is a basic extract of NTS. Use a custom named list, e.g. basic_extract}
}
\value{
A tibble of NTS tables joined together
}
\description{
Build the National Travel Survey data set from a list
}
